fix: remove all dependency cycles
