+++ .eslintrc.js
@@ -22,12 +22,14 @@ module.exports = {
     "sort-keys-fix",
     "typescript-sort-keys",
     "prefer-arrow",
+    "import",
   ],
   rules: {
     "@typescript-eslint/sort-type-union-intersection-members": "error",
     camelcase: "off",
     // "func-style": ["error", "expression", { allowArrowFunctions: true }],
     "func-style": "off",
+    "import/no-cycle": ["error", { maxDepth: 10 }],
     "simple-import-sort/exports": "error",
     "simple-import-sort/imports": "error",
     "sort-keys-fix/sort-keys-fix": "error",
